work in progress
Bare bones of a profile will improve as we progress.
Bare bones of  a post as well

I've added my SQL statements so we can have our databases organized the same way so when we  insert/query we  all use the same format.

I've add a DAO as well feel free to change any and all just make sure that the origianal functionality is the same so we don't have to rework later.