search and file upload on profile
The SearchResults.xhtml links are still being sorted out so it is not working.

A template for file upload is on the upload.xhtml it uses the mediaController.

profile_1.xhtml is the one I am adjusting to send people to from the search results. right now the profile image is linke to a static id but will change as I get it sorted out.